Add JUMPER-SOLDER-CLOSED part, which uses a new layer to connect the two pins (tJumper/bJumper).
To work correctly, the tJumper/bJumper layers must be included in the top/bottom copper gerber layers.

We had to manually edit the library file to specify that these layers use layer numbers 53/54, which is one of several layer pairs that automatically mirror when moving the part between the top/bottom layers. Discovered that from:

    https://forums.autodesk.com/t5/eagle-forum/custom-layer-automatic-mirror/td-p/9937860

    | Built-in there are 4 pairs of layers available to users that automatically mirror each other.
    | These are the 4 pairs 53/54, 55/56, 57/58 and 59/60.
    |
    | Those are the only pairs of layers that have this property.